requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
bootstrap4==0.1.0
Django==5.2
django-bootstrap4==25.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
PyPDF2>=3.0.0
pdf2image>=1.16.0
pillow>=9.0.0
pytesseract>=0.3.10
